AlphaGenome_BatchProcess
This code illustrates one way to process DNA variant data in large quantities. The function extracts variant data from an excel sheet and uses it to run RNA seq predictions. Then, the predicted results are further refined and filtered. Finally, it returns an output matrix and write it to a file named "Variant_Score_Results.xlsx"